Feat: User check to Add, Edit, Remove Api
- Add Email check to Add, Edit, Remove api
- Remove add submit alert
- Refetore Detail db connection
- Change edit api method post to get
- Create error page
- Update List Animation timing (faster)
- Show login user name
- Update router after delete api to button.js
- Add router after add api to add api
- Add MongoDBAdapter
- Add cSpell.words

+++ pages/api/detail/delete/[id].js
@@ -1,12 +1,30 @@
 import { connectDB } from "@/util/database";
 import { ObjectId } from "mongodb";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../../auth/[...nextauth]";
 
 export default async function handler(req, res) {
-  if (req.method === "DELETE") {
+  if (session && req.method === "DELETE") {
+    const session = await getServerSession(req, res, authOptions);
+    const userEmail = session.user.email;
+
     const db = (await connectDB).db("dongflix");
     const movie = await db
       .collection("movies")
-      .deleteOne({ _id: new ObjectId(req.query.id) });
-    res.status(200);
+      .findOne({ _id: new ObjectId(req.query.id) });
+
+    if (movie.author === userEmail) {
+      res.redirect(302, "/");
+      const db = (await connectDB).db("dongflix");
+      const movie = await db
+        .collection("movies")
+        .deleteOne({ _id: new ObjectId(req.query.id) });
+    } else {
+      const errorMessage = "DeleteNotAuthor";
+      res.redirect(302, `/error/?code=${errorMessage}`);
+    }
+  } else {
+    const errorMessage = "guest";
+    res.redirect(302, `/error/?code=${errorMessage}`);
   }
 }
